Create Tic-Tac_toe game
3X3 Numeric Tic-Tac-Toe (Use numbers 1 to 9 instead of X’s and O’s)

One player plays with the odd numbers (1, 3, 5, 7, 9) and the other player plays with the even numbers (2,4,6,8). All numbers can be used only once. The player who puts down 15 points in a line wins (sum of 3 numbers). Always Player with odd numbers starts the game. Once a line contains two numbers whose sum is 15 or greater, there is no way to complete that line, although filling in the remaining cells might be necessary to complete a different line.

Note – Line can be horizontal, vertical or diagonal

Constraints:
1<=Position<=9
1<=Number<=9

Terminal:
Print ‘Welcome to the Game!’.    
Print whether it is Player 1’s or Player 2’s chance.
Get the position and number to be entered from the user.
Show tic tac toe with data.
Continue till the game gets draw or some player wins and show the result.
Ask the user whether to continue for the next game or exit.