Create Kivy-Bank user interface_python file
Requirements:
The project must have the capability to handle large sets of data
A simple method to enroll in various banks with a single account number is the core idea behind this project
Account details like account number and the name of the account holder remain same across all the banks but the details of the bank like IFSC Code, Bank name, MICR code etc. changes according to the bank being considered. (Again, the details of the banks and the account details to be entered into the database can be of user choice) 
A person entering into the portal should be given options to choose an operation that is to be performed related to a particular bank out of all multiple existing banks(only one account pertaining to a single type of bank) 
For example, withdrawal is one such operation and  that should be specific to each and every bank 
All details must be shared between the various accounts but some confidential details cannot be shared. They are as follows: (Two of ‘n’ such confidential parameters have been mentioned below. User should consider these three parameters at least while implementation and allow maximum number of parameters to his own discretion)
Amount in a particular bank account
Transition history of  that bank account
As Project is in the starting phase, we assume any account holder has only accounts in three different banks namely SBI, HDFC, ICICI

UI Requirements:
First screen should be attractive with an appropriate logo and background image for the unified bank portal providing two options for employee and customer
Customer screen, if selected
Should give two widgets namely login, create new  account 
Login, if selected
Should ask for username and password 
Implement forgot password  
Take username, account number and generate a 4-digit OTP after submitting the entered details and send the generated OTP to telegram bot
Enter this OTP and after matched, a window asking for new password and confirm password should pop-up
After entering these details, the password of the particular customer should get updated in the database and the window should go back to the customer login screen
After entering correct credentials, four options must be displayed, But first customer must select the BANK
Delete Account from Bank
Ask for Bank 
Password
Random 4 digit system generated code received on telegram bot
Confirmation message on telegram bot “Account Deleted from  <Bank name>”
Check current balance
Last five transactions
Money transfer
Account number to be transferred to must be entered and after entering, name and mobile number related to that account must be displayed
Then amount must be entered 
Message should go to telegram bot displaying the amount sent and the account number to which the money has been sent with Time and Date
After selecting the “New User Registration” following details must be shown on the screen
User must enter following details for successful registration
Name 
Date of Birth
Address
Country
Bank (To which the customer  want to be registered to)
Mobile Number
After click on the submit button a 7 digit number must be generated and to be sent to the User Telegram Bot And a pop window generated with message “Successfully Registered” and direct the user to login screen
eg. You are successfully registered for the <Bank Name> with account number “<10_digit_number>”. 
EG. “2345362768”
“Employee login screen”
User-Id (Must be in the form of “SBI234” and employee already registered in the bank )
Password 


After successful login, An employee can be able to perform the following tasks

Update account details of customer, based on the details filled up by the customer 
The following details can be updated  
Address
Mobile Number
Date of Birth
When any data related to customer gets updated, an Acknowledge  message  must be sent to customer via telegram bot with Time and Date
Employee can see the account details( Details related to employee of one of the multiple banks) of all customers

Note - The evaluation will be done based on execution of the above tasks and visual aesthetics of the application. So, you are free to imagine your idea and create user interface.